requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
colorama>=0.4.6
numpy>=1.24.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
dateutils>=0.6.12
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
